Add files via upload
Currently, I am done by creating the service layer capable of communicating with the API layer.

Run the endpoint: http://localhost:8080/api/v1/student
